fix: Move dir, fix paths, options
