+++ dags/ethereum_classic_load_dag.py
@@ -4,7 +4,7 @@ import logging
 
 from airflow.models import Variable
 
-from ethereumetl.build_load_dag import build_load_dag
+from build_load_dag import build_load_dag
 
 logging.basicConfig()
 logging.getLogger().setLevel(logging.DEBUG)
